# Reduce VTGate Normalizer multiple AST walks 