<!--
  Thank you for your contribution to the Vitess project.
  How to contribute: https://vitess.io/docs/contributing/
  Please first make sure there is an open Issue to discuss the feature/fix suggested in this PR.
  If this is a new feature, please mark the Issue as "RFC".
 -->

<!-- if this PR is Work in Progress please create it as a Draft Pull Request -->

## Description

<!-- A few sentences describing the overall goals of the pull request's commits. -->
<!-- If this is a bug fix and you think the fix should be backported, please write so. -->

This PR reduces the ast walk in VTGate normalizer code.

Benchmark shows `~5% normalizer` improvement
```
goos: darwin
goarch: arm64
pkg: vitess.io/vitess/go/vt/sqlparser
cpu: Apple M1 Max
                                     │ /Users/planetscale/benchmark/v2.txt │ /Users/planetscale/benchmark/v3.txt │
                                     │               sec/op                │    sec/op     vs base               │
Normalize-4                                                    1.217µ ± 5%    1.103µ ± 1%   -9.41% (p=0.002 n=6)
NormalizeTraces/django_queries.txt-4                           185.9µ ± 2%    167.3µ ± 1%  -10.04% (p=0.002 n=6)
NormalizeTraces/lobsters.sql.gz-4                              9.613m ± 3%    8.348m ± 2%  -13.16% (p=0.002 n=6)
NormalizeVTGate-4                                              71.89m ± 1%    70.74m ± 1%   -1.60% (p=0.041 n=6)
NormalizeTPCCBinds-4                                           18.04µ ± 1%    17.97µ ± 1%        ~ (p=0.093 n=6)
NormalizeTPCCInsert-4                                          63.65m ± 2%    63.75m ± 1%        ~ (p=0.937 n=6)
NormalizeTPCC-4                                                51.59m ± 1%    50.73m ± 2%   -1.67% (p=0.026 n=6)
geomean                                                        1.374m         1.302m        -5.30%

                                     │ /Users/planetscale/benchmark/v2.txt │ /Users/planetscale/benchmark/v3.txt │
                                     │                B/op                 │     B/op      vs base               │
NormalizeTraces/django_queries.txt-4                          178.3Ki ± 0%   167.1Ki ± 0%   -6.28% (p=0.002 n=6)
NormalizeTraces/lobsters.sql.gz-4                             7.072Mi ± 0%   6.277Mi ± 0%  -11.25% (p=0.002 n=6)
NormalizeVTGate-4                                             48.27Mi ± 0%   47.46Mi ± 0%   -1.67% (p=0.002 n=6)
NormalizeTPCCBinds-4                                          11.35Ki ± 0%   11.36Ki ± 0%   +0.13% (p=0.002 n=6)
NormalizeTPCCInsert-4                                         50.63Mi ± 0%   50.65Mi ± 0%   +0.04% (p=0.002 n=6)
NormalizeTPCC-4                                               35.77Mi ± 0%   35.40Mi ± 0%   -1.02% (p=0.002 n=6)
geomean                                                       3.257Mi        3.145Mi        -3.43%

                                     │ /Users/planetscale/benchmark/v2.txt │ /Users/planetscale/benchmark/v3.txt  │
                                     │              allocs/op              │  allocs/op   vs base                 │
NormalizeTraces/django_queries.txt-4                           6.025k ± 0%   5.389k ± 0%  -10.56% (p=0.002 n=6)
NormalizeTraces/lobsters.sql.gz-4                              252.3k ± 0%   213.2k ± 0%  -15.49% (p=0.002 n=6)
NormalizeVTGate-4                                              1.112M ± 0%   1.073M ± 0%   -3.51% (p=0.002 n=6)
NormalizeTPCCBinds-4                                            301.0 ± 0%    301.0 ± 0%        ~ (p=1.000 n=6) ¹
NormalizeTPCCInsert-4                                          958.0k ± 0%   958.0k ± 0%        ~ (p=0.671 n=6)
NormalizeTPCC-4                                                836.9k ± 0%   818.5k ± 0%   -2.20% (p=0.002 n=6)
geomean                                                        86.11k        81.40k        -5.48%
¹ all samples are equal

```

## Related Issue(s)

<!-- List related issues and pull requests. If this PR fixes an issue, please add it using Fixes #????  -->

## Checklist

-   [x] Tests were added or are not required
-   [x] Did the new or modified tests pass consistently locally and on CI?
-   [x] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->